Compare Files Using Unified Diffs

unidiff() is a function designed to generate unified diffs between two files or text inputs. Unified diffs are widely used in version control systems to highlight differences between two versions of a file in a compact and readable format.

unidiff(
  old,
  new,
  create_head = TRUE,
  with_context = FALSE,
  context_length = 3,
  ignore_whitespace = NULL,
  algorithm = "minimal",
  indent_heuristic = FALSE
)

Arguments

old

The original file or text input. This can be a file path, a URL, or a string of text.

new

The modified file or text input to compare against the original. This can also be a file path, a URL, or a string of text.

create_head

A logical flag (default TRUE) indicating whether to include a header in the diff output. The header contains the filenames or identifiers of the files being compared.

with_context

A logical flag (default FALSE) specifying whether to include the full context in the diff output. Including context helps to provide a more complete view of the changes.

context_length

An integer value (default 3) that determines the number of lines of context to include around each change. This is only relevant if with_context is TRUE.

ignore_whitespace

A character vector indicating how whitespace differences should be treated. Options include "all", "change", "at_eol", "cr_at_eol", and "blank_lines", allowing for fine-grained control over which whitespace differences to ignore.

algorithm

A string specifying the diff algorithm to use. Options are "minimal", "patience", and "histogram", with "minimal" being the default. Each algorithm has different characteristics suited to various types of text comparisons.

indent_heuristic

A logical flag (default FALSE) that, when enabled, applies an additional heuristic to better handle indented lines, improving the accuracy of the diff in some cases.

Value

A character string containing the unified diff.

Details

The unidiff() function compares two inputs—whether they are files, URLs, or raw text—and generates a unified diff that highlights the differences between them. It supports advanced features like customizable context, whitespace handling, and different diff algorithms to suit various needs. The function can also handle inputs from URLs, making it flexible for remote file comparisons. The output is a unified diff string, which can be directly used in version control systems or for manual inspection of changes. For identical files, no diff is returned, and an empty character string is provided.

Examples

cat(unidiff(old = "foo bar", new = "foo baz"))
#> ===================================================================
#> --- old
#> +++ new
#> @@ -1 +1 @@
#> -foo bar
#> \ No newline at end of file
#> +foo baz
#> \ No newline at end of file